GameSessionKey used instead of long session identifier

+++ server/player/player-common/src/main/resources/sql/schema/h2/player-schema.sql
@@ -10,6 +10,6 @@
         IMAGE_URL varchar(255),
         LAST_NAME varchar(64),
         NICK_NAME varchar(64),
-        VERSION bigint,
+        VERSION integer,
         primary key (PLAYER_ID)
     );
